Todo-Application
A to-do list is a simple organizational tool used to keep track of tasks that need to be completed. It typically consists of a list of items or activities that a person needs or wants to accomplish within a certain timeframe. To-do lists can be written on paper, on a digital device, or in specialized apps designed for task management. They are helpful for prioritizing tasks, managing time effectively, and reducing the feeling of being overwhelmed by providing a clear overview of what needs to be done.